Extract delete doctor dialog
